refactor frontend, cleanup desktop view
