+++ src/App.js
@@ -7,9 +7,9 @@ import { useAuth } from './contexts/auth';
 import Header from './components/Header';
 import PrivateRoute from './components/PrivateRoute';
 import Welcome from './components/Welcome';
-import Dashboard from './components/dashboardComponents/Dashboard';
 import Loading from './components/Loading';
 import Mobile from './components/mobile/Mobile';
+import MobileDashboard from './components/mobile/MobileDashboard.js';
 
 // function initializeAnalytics() {
 //   ReactGA.initialize('UA-157827018-1');
@@ -19,28 +19,7 @@ import Mobile from './components/mobile/Mobile';
 
 function App() {
   const { googleApi } = useAuth();
-  //mobile breakpoint
-  const breakPoint = 768
-
-    //state for current window dimensions
-    const [dimensions, setDimensions] = useState({ 
-      height: window.innerHeight,
-      width: window.innerWidth
-    })
-    //listens for window resize to set above state. event listener must be cleaned up in the return as below or will cause bugs.
-    useEffect(() => {
-      function handleResize() {
-        setDimensions({
-          height: window.innerHeight,
-          width: window.innerWidth
-        })
-      }
-      window.addEventListener('resize', handleResize)
-      return _ => {
-        window.removeEventListener('resize', handleResize)
-      }
-      },[])
-
+  
   //holds the current user, mostly to stop problems with user's profile image from not populating correctly
   const [userState, setUserState] = useState({})
   let currentUser = googleApi.currentUser
@@ -54,47 +33,6 @@ function App() {
     return <Loading />;
   }
 
-  //deals with desktop vs. mobile and logged in vs not. could probably be consolidated
-  
-  //if at desktop size
-  if(dimensions.width > breakPoint){
-
-      //if user is logged in, send them to the dashboard
-      if (googleApi.currentUser){
-        return (
-          <Stack pos="relative" w="100%" minHeight="100vh">
-            <Header userState={userState} />
-            <Route path="/">
-              <Authenticate />
-            </Route>
-            {googleApi.currentUser && <Route exact path="/">
-              <Welcome />
-            </Route>}
-            <PrivateRoute path="/:id/dashboard">
-              <Dashboard setUserState={setUserState}/>
-            </PrivateRoute>
-          </Stack>
-        );
-      //if not logged in, attempt to authenticate them
-      } else {
-        return (
-          <Stack pos="relative" w="100%" minHeight="100vh">
-            <Header />
-            <Route path="/authenticate/google">
-              <Authenticate />
-            </Route>
-            <Route exact path="/">
-              <Welcome />
-            </Route>
-            <PrivateRoute path="/:id/dashboard">
-              <Dashboard />
-            </PrivateRoute>
-          </Stack>
-
-        )
-      }
-//if at mobile size
-} else if(dimensions.width <= breakPoint){
   //if user is logged in
   if(googleApi.currentUser){
     return (
@@ -124,12 +62,11 @@ function App() {
                 <Welcome />
               </Route>
               <PrivateRoute path="/:id/dashboard">
-                <Dashboard />
+                <MobileDashboard />
               </PrivateRoute>
             </Stack>
     )
   }
 
-}
 }
 export default App;
