--- src/components/dashboardComponents/CreateTemplateForm.js
@@ -1,63 +1,0 @@
-import React, {useContext} from 'react';
-import {DashboardContext} from '../../contexts/DesktopContexts'
-import { useForm } from 'react-hook-form';
-import { Button, Input } from '@chakra-ui/core';
-import {addTemplate} from '../../utils/helperFunctions'
-
-const CreateTemplateForm = () => {
-
-  const {
-    formOpen,
-    setFormOpen,
-    setTemplateList,
-    currentUser} = useContext(DashboardContext);
-
-    
-  const { register, handleSubmit } = useForm();
-
-  // Submit for template form
-const onSubmit = async formData => {
-    const template = addTemplate(formData, currentUser);
-    await setTemplateList(prevTemplates => [...prevTemplates, template]);
-    setFormOpen(!formOpen);
-  };
-
-  return (
-
-    <div className="Form">
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <Input
-          type="text"
-          placeholder="title"
-          name="title"
-          id="title"
-          ref={register({ maxLength: 80, required: true })}
-        />
-        <Input
-          type="text"
-          placeholder="notes"
-          name="notes"
-          id="notes"
-          ref={register({ maxLength: 100 })}
-        />
-        <Input
-          type="time"
-          name="starttime"
-          id="starttime"
-          ref={register({ required: true })}
-        />
-        <Input
-          type="time"
-          name="endtime"
-          id="endtime"
-          ref={register({ required: true })}
-        />
-
-        <Button type="submit">Submit</Button>
-      </form>
-    </div>
-
-  );
-};
-
-export default CreateTemplateForm;
