--- src/components/dashboardComponents/calendarComponents/EventsModal.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Modal from '@material-ui/core/Modal';
-
-const EventModal = () => {
-    return (
-        <div>
-            <Modal>
-                
-            </Modal>
-        </div>
-    )
-}
-
-export default EventModal;
