--- src/contexts/DesktopContexts.js
@@ -1,4 +1,0 @@
-import {createContext} from 'react'
-
-export const DesktopContext = createContext();
-export const DashboardContext = createContext();
